Add gist selector option to gist edit command
