+++ pkg/cmd/gist/shared/shared.go
@@ -5,10 +5,15 @@ import (
 	"fmt"
 	"net/http"
 	"net/url"
+	"sort"
 	"strings"
 	"time"
 
+	"github.com/AlecAivazis/survey/v2"
 	"github.com/cli/cli/v2/api"
+	"github.com/cli/cli/v2/internal/text"
+	"github.com/cli/cli/v2/pkg/iostreams"
+	"github.com/cli/cli/v2/pkg/prompt"
 	"github.com/gabriel-vasile/mimetype"
 	"github.com/shurcooL/githubv4"
 )
@@ -170,4 +175,56 @@ func IsBinaryContents(contents []byte) bool {
 		}
 	}
 	return isBinary
+}
+
+func PromptGists(client *http.Client, host string, cs *iostreams.ColorScheme) (gistID string, err error) {
+	gists, err := ListGists(client, host, 10, "all")
+	if err != nil {
+		return "", err
+	}
+
+	if len(gists) == 0 {
+		return "", nil
+	}
+
+	var opts []string
+	var result int
+	var gistIDs = make([]string, len(gists))
+
+	for i, gist := range gists {
+		gistIDs[i] = gist.ID
+		description := ""
+		gistName := ""
+
+		if gist.Description != "" {
+			description = gist.Description
+		}
+
+		filenames := make([]string, 0, len(gist.Files))
+		for fn := range gist.Files {
+			filenames = append(filenames, fn)
+		}
+		sort.Strings(filenames)
+		gistName = filenames[0]
+
+		gistTime := text.FuzzyAgo(time.Now(), gist.UpdatedAt)
+		// TODO: support dynamic maxWidth
+		description = text.Truncate(100, text.RemoveExcessiveWhitespace(description))
+		opt := fmt.Sprintf("%s %s %s", cs.Bold(gistName), description, cs.Gray(gistTime))
+		opts = append(opts, opt)
+	}
+
+	questions := &survey.Select{
+		Message: "Select a gist",
+		Options: opts,
+	}
+
+	//nolint:staticcheck // SA1019: prompt.SurveyAskOne is deprecated: use Prompter
+	err = prompt.SurveyAskOne(questions, &result)
+
+	if err != nil {
+		return "", err
+	}
+
+	return gistIDs[result], nil
 }
